requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
ddt==1.2.1
Deprecated==1.2.6
gitdb2==2.0.5
GitPython==3.0.1
idna==2.8
PyGithub==1.43.8
PyJWT==1.7.1
requests==2.22.0
smmap2==2.0.5
urllib3==1.25.3
wrapt==1.11.2
